Fix parity write returning negative value
The metric we get is still off when compared to Python, so we'd need to
look into that, but the calculation should be correct. The problem was
due to calculation error in both data_center and gc_stripe function.
Further alternation may be required for the derived gc strategy as well.

Also changed all of the return metric to double for now. Although double
values take up more space and flops take longer, 1) it is kind of
required when considering the noise value and 2) amount of calculations
we perform on these metric values is not nearly enough for this to
ultimately matter. From brief testing, it looks like the simulator
finishes roughly around same time (which is still miles faster than
python).